refactor: give the channel contract its own home, and serving its own file
