+++ src/channels/agentcore.ts
@@ -30,8 +30,8 @@ import { timingSafeEqual } from "node:crypto";
 import type { Agent } from "../agent.ts";
 import type { StateSync, StateUrls } from "./agentcore-state.ts";
 import { beginWork, onIdle } from "./busy.ts";
-import type { ChannelHandler, Routes } from "../host/node.ts";
-import { router } from "../host/node.ts";
+import type { ChannelHandler, Routes } from "../channel.ts";
+import { router } from "../channels/serve.ts";
 import { log } from "../log.ts";
 import { rememberWakeAlarmUrl } from "../schedule/wake-alarm.ts";
 import type { ScheduleFireOutcome } from "../schedule/scheduler.ts";
