+++ src/channels/lark/lark.ts
@@ -3,7 +3,7 @@
  * this module binds it to Lark's cloud profile and exposes natural Lark-branded public names. Lark's
  * weaker control-plane capabilities live in onboarding/registration, not in a fork of the turn engine.
  */
-import type { ChannelModule, LongConnectionChannelModule } from "../../host/node.ts";
+import type { ChannelModule, LongConnectionChannelModule } from "../../channel.ts";
 import { LARK_COMPAT_CLOUD } from "../feishu/cloud.ts";
 import {
   type FeishuChannelOptions,
