+++ src/cli/serve.ts
@@ -14,7 +14,8 @@ import { text } from "../channels/respond.ts";
 import { type LoadedLongConnectionChannel, loadChannels } from "../engines/pi/channel.ts";
 import { reportModuleLoadFailures } from "../engines/pi/report.ts";
 import { answersLocalhost, bindLabel, classifyBind, clientHost } from "../bind.ts";
-import { type Routes, parseRouteKey, routePathsOverlap, router, serveNode } from "../host/node.ts";
+import type { Routes } from "../channel.ts";
+import { parseRouteKey, routePathsOverlap, router, serveNode } from "../channels/serve.ts";
 import { log } from "../log.ts";
 import { openExternalUrl } from "../open-url.ts";
 import { loadSchedules } from "../schedule/discover.ts";
